+++ resip/stack/ExtensionParameter.cxx
@@ -14,7 +14,7 @@ ExtensionParameter::ExtensionParameter(const Data& name)
    : mName(name)
 {
    assert(!mName.empty());
-   assert(ParameterTypes::getType(mName.data(), mName.size()) == ParameterTypes::UNKNOWN);
+   assert(ParameterTypes::getType(mName.data(), (unsigned int)mName.size()) == ParameterTypes::UNKNOWN);
 }
 
 const Data&
